added full firebase-support affair part GRATZ!

+++ app/src/main/java/com/bytebuilding/affairmanager/activities/DetailAffairActivity.java
@@ -18,34 +18,49 @@ public class DetailAffairActivity extends AppCompatActivity {
 
     @BindView(R.id.detailActivity_title)
     TextView title;
+
     @BindView(R.id.detailActivity_title_dynamic)
     TextView titleDynamic;
+
     @BindView(R.id.detailActivity_description)
     TextView description;
+
     @BindView(R.id.detailActivity_description_dynamic)
     TextView descriptionDynamic;
+
     @BindView(R.id.detailActivity_date)
     TextView date;
+
     @BindView(R.id.detailActivity_date_dynamic)
     TextView dateDynamic;
+
     @BindView(R.id.detailActivity_time)
     TextView time;
+
     @BindView(R.id.detailActivity_time_dynamic)
     TextView timeDynamic;
+
     @BindView(R.id.detailActivity_object)
     TextView object;
+
     @BindView(R.id.detailActivity_object_dynamic)
     TextView objectDynamic;
+
     @BindView(R.id.detailActivity_type)
     TextView type;
+
     @BindView(R.id.detailActivity_type_dynamic)
     TextView typeDynamic;
+
     @BindView(R.id.detailActivity_place)
     TextView place;
+
     @BindView(R.id.detailActivity_place_dynamic)
     TextView placeDynamic;
+
     @BindView(R.id.detailActivity_status)
     TextView status;
+
     @BindView(R.id.detailActivity_status_dynamic)
     TextView statusDynamic;
 
@@ -74,7 +89,7 @@ public class DetailAffairActivity extends AppCompatActivity {
         objectDynamic.setText(outerBundle.getString("object"));
         typeDynamic.setText(outerBundle.getString("type"));
         placeDynamic.setText(outerBundle.getString("place"));
-        if (outerBundle.getInt("status") == 2) {
+        if (outerBundle.getLong("status") == 2) {
             statusDynamic.setText(R.string.detailActivity_status_done);
         } else {
             statusDynamic.setText(R.string.detailActivity_status_current);
